ManuelBastioniLab: Heuristics
An attempt to get smart while creating the lipsync.
Basically there are a set of heuristics that I assume are generally
true when creating lip sync. Ex: The mouth closes after a word if the
number of frames between the last word and the next is greater than
X number of frames, and so on.

Signed-off-by: Amir Shehata <amir.shehata@gmail.com>