ManuelBastioniLab: Heuristic adjustments and fixes
Created another heuristic function which averages the value of the
keyframes of the bone across a window specified by the user. The window
size is in number of keyframes. But it can be given as a float.
For example: 1.35 will mean that the window size is 1 before and 1 after
the current key frame. But when calculating the average we'll divide
by 1.35. Also we never drop the value of the current key frame beyond
a 3rd of its original value

Signed-off-by: Amir shehata <amir.shehata@gmail.com>